fix: move InsightType to robotoff.types
