--- robotoff/insights/__init__.py
@@ -1,3 +1,0 @@
-from .dataclass import InsightType
-
-__all__ = ["InsightType"]
